supporting-blog-content/elasticsearch-through-apache-kafka/requirements.txt
kafka-python==2.0.2
elasticsearch==7.10.0